Fix compile errors

+++ r2dbc-migrate-core/src/test/java/name/nkonev/r2dbc/migrate/core/MssqlTestcontainersTest.java
@@ -160,7 +160,7 @@ public class MssqlTestcontainersTest {
 
         ConnectionFactory connectionFactory = makeConnectionMono(mappedPort);
 
-        Mono<Integer> integerMono = Mono.usingWhen(
+        Mono<Long> integerMono = Mono.usingWhen(
             connectionFactory.create(),
             connection -> Mono
                 .from(connection.createStatement("create schema \"my scheme\"").execute())
